Bump Reffy to v15 and adjust code
The only breaking change that affects Web IDL validation is the separation
between the `idl` and `idlparsed` modules in v6.0.0, which made the extracted
IDL available under a different path in the result.

A couple of minor improvements have been made to IDL extraction but no other
substantive change that would require updating the code.